* src/header.c: use size_t for header_size. (skip_msdos_sfx1_code): rename to `seek_lha_header'.
* src/lha.h: use size_t for header_size.

* src/lhadd.c (cmd_add): rename to `seek_lha_header'.
(cmd_delete): ditto.

* src/lhext.c (cmd_extract): ditto.

* src/lhlist.c (cmd_list): ditto.

* src/prototypes.h: updated.

* tests/lha-test15: added tests to search any level headers.


git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/lha/lha/trunk@591 6a8cc165-1e22-0410-a132-eb4e3f353aba

+++ src/lha.h
@@ -222,7 +222,7 @@ struct string_pool {
 };
 
 typedef struct LzHeader {
-    long            header_size;
+    size_t          header_size;
     int             size_field_length;
     char            method[METHOD_TYPE_STORAGE];
     long            packed_size;
